+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # CHANGELOG
 
+## [v0.2.0][v0.2.0] - May 2020
+
+- Added component `<symbol-transaction-icon>`
+- Added `identity-obj-proxy` for jest image module wrapping
+- Added `@rollup/plugin-image` for bundling images as base64 URIs (self-contained library)
+- Removed www test service worker and cleaned dependencies
+- Added multiple output targets configuration for better compatibility
+
 ## [v0.1.0][v0.1.0] - May 2020
 
 - Added component `<symbol-node-health-icon>`
 - Added configuration for StencilJS webcomponents compiler
 
+[v0.2.0]: https://github.com/symbol/symbol-components/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/symbol/symbol-components/releases/tag/v0.1.0